requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bitarray==2.9.2
boltons==21.0.0
bracex==2.5.post1
cbor2==5.6.5
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
click-option-group==0.5.6
colorama==0.4.6
coverage==7.6.4
crytic-compile==0.3.7
cytoolz==0.12.3
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
eth-account==0.11.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.2.2
face==22.0.0
frozenlist==1.4.1
glom==22.1.0
googleapis-common-protos==1.65.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==7.1.0
iniconfig==2.0.0
jiter==0.6.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
openai==1.52.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
packaging==24.1
parsimonious==0.10.0
peewee==3.17.6
pluggy==1.5.0
prettytable==3.11.0
protobuf==4.25.5
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.5.3
rlp==4.0.1
rpds-py==0.20.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
semgrep==1.93.0
setuptools==75.1.0
slither-analyzer==0.10.4
sniffio==1.3.1
solc-select==1.0.4
tomli==2.0.2
toolz==0.12.1
tqdm==4.66.5
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
wcmatch==8.5.2
wcwidth==0.2.13
web3==6.20.3
websockets==13.1
wrapt==1.16.0
yarl==1.13.1
zipp==3.20.2
